requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
Pillow==7.1.0
pytz==2019.3
sqlparse==0.3.1
